move Plans field from Csar to HashedFile refs #301
