+++ src/main/java/org/opentosca/csarrepo/model/HashedFile.java
@@ -1,7 +1,9 @@
 package org.opentosca.csarrepo.model;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 import javax.persistence.Column;
@@ -9,6 +11,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.MapKey;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
@@ -40,6 +43,11 @@ public class HashedFile {
 	@OneToMany(mappedBy = "hashedFile")
 	@LazyCollection(LazyCollectionOption.FALSE)
 	private List<CsarFile> csarFiles;
+
+	@OneToMany(mappedBy = "csar")
+	@LazyCollection(LazyCollectionOption.FALSE)
+	@MapKey(name = "name")
+	private Map<String, CsarPlan> csarPlans = new HashMap<String, CsarPlan>();
 
 	public HashedFile() {
 		this.csarFiles = new ArrayList<CsarFile>();
@@ -118,4 +126,24 @@ public class HashedFile {
 	public List<CsarFile> getCsarFiles() {
 		return csarFiles;
 	}
+
+	/**
+	 * Adds a plan entry for a plan
+	 * 
+	 * @param planId
+	 * @param planReference
+	 */
+	public void addPlan(String planId, CsarPlan plan) {
+		this.csarPlans.put(planId, plan);
+	}
+
+	/**
+	 * returns the mappings from planID to planReference (ID -> zipFileName)
+	 * 
+	 * @return Map
+	 */
+	public Map<String, CsarPlan> getPlans() {
+		return this.csarPlans;
+	}
+
 }
